+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptsnippetop.py
@@ -1831,7 +1831,7 @@ class CreatePromptSnippetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-10T15:50:12.206Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-10T17:12:36.462Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")