+++ packages/orq-rc/RELEASES.md
@@ -309,3 +309,13 @@ Based on:
 - [python v3.1.0-rc.53] packages/orq-rc
 ### Releases
 - [PyPI v3.1.0-rc.53] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.53 - packages/orq-rc
+
+## 2025-02-10 17:14:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.487.1 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.1.0-rc.54] packages/orq-rc
+### Releases
+- [PyPI v3.1.0-rc.54] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.54 - packages/orq-rc